pyproject.toml
[project]
name = "linear-api-mcp"
version = "0.1.0"
description = "FastMCP server for Linear API - Query initiatives, projects, and documents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0"
]
